setup NextJS project with Login functionality
In this commit we have introduces NextJS project into the repo and
created a Login page. We had added the Dockerfile to containerize
the application. After that I have updated the README.md file to
include steps to turn the application on via simple and docker
method.

I have also added the reasoning behind my designs decisions took
in the project.